requirements.txt
anthropic
python-dotenv
metaphone
pydantic

